+++ source/frontend/src/modules/login/Login.vue
@@ -1,11 +1,12 @@
 <template lang="html">
   <div id="index">
-    <login-form></login-form>
+    <login-form @login="login"></login-form>
     <axios-interceptor />
   </div>
 </template>
 
 <script type="text/javascript">
+import RSA from "rsajs";
 import LoginForm from "./components/login-form/login-form";
 
 export default {
@@ -22,7 +23,41 @@ export default {
     LoginForm
   },
   watch: {},
-  methods: {}
+  methods: {
+    async login(userInfo){
+
+      let publicKey = await this.getPublicKey();
+      let password = userInfo.password;
+
+      if (publicKey && publicKey !== "") {
+        password = this.encrypt(publicKey, userInfo.password);
+      }
+
+      let xAccessToken = await this.auth(userInfo.username, password);
+      console.info(publicKey, xAccessToken);
+
+      if (xAccessToken && xAccessToken !== "") {
+      window.location.href = "app";
+      }
+    },
+
+    async getPublicKey() {
+      return this.$api.get("/api/authenticate/key");
+    },
+
+    async auth(username, password) {
+      return this.$api.post("/api/authenticate", {
+        username: username,
+        password: password
+      });
+    },
+
+    encrypt(key, source) {
+      let rsa = new RSA();
+      rsa.setKey(key);
+      return rsa.encrypt(source);
+    }
+  }
 };
 </script>
