+++ source/frontend/src/modules/login/components/login-form/login-form.vue
@@ -2,29 +2,30 @@
   <div>
     <form class="form login-form" @submit.prevent="submit">
       <icon data="@icon/user-setting.svg"></icon>
-      <label for="username"> 사용자명1:</label><br />
+      <label for="username"> 사용자명1:</label><br/>
       <input
-        class="text-input"
-        type="text"
-        id="username"
-        v-model="username"
-        maxlength="10"
-        required
-      /><br />
+          class="text-input"
+          type="text"
+          id="username"
+          v-model="username"
+          maxlength="10"
+          required
+      /><br/>
       <icon data="@icon/key.svg"></icon>
-      <label for="password">비밀번호:</label><br />
+      <label for="password">비밀번호:</label><br/>
       <input
-        class="text-input"
-        type="password"
-        id="password"
-        v-model="password"
-        maxlength="15"
-        required
-      /><br />
+          class="text-input"
+          type="password"
+          id="password"
+          v-model="password"
+          maxlength="15"
+          required
+      /><br/>
       <select id="locale" class="login_form-locale">
         <option value="korean">한국어</option>
-        <option value="English">English</option></select
-      ><br />
+        <option value="English">English</option>
+      </select
+      ><br/>
       <button type="submit" class="button button--primary button--lg w-12_12">
         로그인
       </button>
@@ -33,7 +34,6 @@
 </template>
 
 <script type="text/javascript">
-import RSA from "rsajs";
 
 export default {
   name: "LoginForm",
@@ -49,34 +49,13 @@ export default {
   components: {},
   watch: {},
   methods: {
-    async submit() {
-      let publicKey = await this.getPublicKey();
-      let password = this.password;
-
-      if (publicKey && publicKey !== "") {
-        password = this.encrypt(publicKey, this.password);
+    submit() {
+      if (this.username && this.password) {
+        this.$emit("login", {
+          username: this.username,
+          password: this.password
+        })
       }
-
-      let xAccessToken = await this.auth(this.username, password);
-      console.info(publicKey, xAccessToken);
-
-      if (xAccessToken && xAccessToken !== "") {
-        window.location.href = "app";
-      }
-    },
-    async getPublicKey() {
-      return this.$api.get("/api/authenticate/key");
-    },
-    async auth(username, password) {
-      return this.$api.post("/api/authenticate", {
-        username: username,
-        password: password
-      });
-    },
-    encrypt(key, source) {
-      let rsa = new RSA();
-      rsa.setKey(key);
-      return rsa.encrypt(source);
     }
   }
 };
